feat: add http clients for metadata services
